abstracted the property table for better readability
